Imports cleanup (#828)
Changes around imports; I was doing some refactoring and found myself having to count how many ../../ I needed to type ... Not anymore

This is my first crack at it and definitely open to suggestions on refining this

Changes: Backend and Frontend
Applying import/order eslint rule. This is to help keep our import statements in a sane order and will report an eslint error if they aren't. Luckily it's something eslint can autofix. The order configured in this PR:
built-in modules
external modules
internal modules
parent/sibling imports
everything else
Changes: Only on the Frontend
Because Next.js supports tsconfig paths aliases out of the box, I have updated packages/front-end/tsconfig.json with a paths alias of @/ to alias the root dir of front-end. This will allow us to refer to project directories with absolute paths. For example, ../../services/utils becomes @/services/utils
I ran a codemod to transform the code. Can publish that script in a separate gist later
Open to suggestions on alias naming. I opted for @/ because without it it may be hard to notice at a glance which imports are external libraries and which are ours
I only applied the path aliasing changes to the frontend because it would require a little more work on the backend, including a change in how we boot the production node app. I would like to socialize these changes first and make sure we iron out the kinks before making the more precarious change on the backend.