Bug fix when user's name is undefined (#1134)
This was breaking our EventModel validation.

This is a quick bug fix to unblock affected users. For a proper fix, we should apply a migration to the UserModel and ensure name is always defined. This requires refactoring the UserModel to only export methods and use a toInferface method, so it's a little bit bigger scope.