Support BigQuery Information Schemas & Auto Metrics with multiple data sets (#1585)
Removes requirement for a BigQuery data source to have a default dataset. Now, this will run multiple queries - 1 to get all of the datasets, and then run a query for each dataset to build the informationSchema. Also adds provisions to support a new frontend UI when building auto metrics to provide a drop-down that allows the user to select the datasource they want to generate metrics from.