Handle race conditions in the feature toggle endpoint (#1848)
For example, if two people have the feature page open and hit the kill switch at the same time. It worked fine previously, but would log 2 separate audit events.  The first one would show the proper diff, the second one would be an empty diff (e.g. changing `false` to `false`).

This PR adds an early return if the feature is already in the desired state so we avoid adding that 2nd empty audit entry.